requirements.txt
# --- Discord 봇 핵심 라이브러리 ---
discord.py>=2.3.0

# --- AI 및 데이터 처리 ---
# Google Gemini AI 라이브러리
google-generativeai>=0.5.0

# 벡터 및 행렬 연산을 위한 라이브러리 (RAG 유사도 계산용)
numpy>=1.26.0

# 비동기 SQLite 드라이버 (메인 봇 DB 연결용)
aiosqlite>=0.19.0

# --- 한국어 자연어 처리 (NLP) ---
# (konlpy, mecab-ko 등은 현재 버전에서 사용되지 않음)

# --- 기타 유틸리티 ---
# .env 파일을 읽기 위한 라이브러리
python-dotenv>=1.0.0

# 시간대 처리를 위한 라이브러리
pytz

# HTTP 요청을 위한 라이브러리 (기상청 API 등)
requests

# 비동기 HTTP 요청을 위한 라이브러리
aiohttp>=3.9.0
aiohttp-retry>=2.8.0